+++ packages/expo-router/src/native-tabs/types.ts
@@ -12,10 +12,9 @@ import type { SFSymbol } from 'sf-symbols-typescript';
 import type {
   DefaultRouterOptions,
   ParamListBase,
-  RouteProp,
-  ScreenListeners,
   TabNavigationState,
 } from '../react-navigation/native';
+import type { StandardUseNavigationBuilderOptions } from '../standard-navigation';
 import type { ScreenProps } from '../useScreens';
 
 /**
@@ -451,11 +450,12 @@ export interface NativeTabsProps extends PropsWithChildren {
    * </NativeTabs>
    * ```
    */
-  screenListeners?:
-    | ScreenListeners<TabNavigationState<ParamListBase>, NativeTabNavigationEventMap>
-    | ((prop: {
-        route: RouteProp<ParamListBase, string>;
-      }) => ScreenListeners<TabNavigationState<ParamListBase>, NativeTabNavigationEventMap>);
+  screenListeners?: StandardUseNavigationBuilderOptions<
+    TabNavigationState<ParamListBase>,
+    object,
+    NativeTabNavigationEventMap
+  >['screenListeners'];
+
   /**
    * Props passed to the underlying native tab host implementation in `react-native-screens`.
    * Use this to configure props that are not directly exposed by Expo Router.